Refactored code into new util file
